requirements.txt
Django>=4.0,<5.0
djangorestframework
gunicorn
psycopg2-binary
django-cors-headers
python-decouple
psycopg2-binary
dj-database-url

